# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blooma API",
        "version": "1.0.0",
        "description": "REST API for Blooma Studio"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Base URL"
        }
    ],
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "__session"
            }
        }
    },
    "security": [
        {
            "cookieAuth": []
        }
    ],
    "tags": [
        {
            "name": "User",
            "description": "User account and credits"
        },
        {
            "name": "Models",
            "description": "Manage character references"
        },
        {
            "name": "Locations",
            "description": "Manage background locations"
        },
        {
            "name": "Camera Settings",
            "description": "Manage custom camera presets"
        },
        {
            "name": "Gallery",
            "description": "Manage generated images"
        }
    ],
    "paths": {
        "/user/credits": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get User Credits",
                "description": "Retrieve current credit balance and subscription status",
                "responses": {
                    "200": {
                        "description": "Credit balance details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "total": {
                                                    "type": "number"
                                                },
                                                "used": {
                                                    "type": "number"
                                                },
                                                "remaining": {
                                                    "type": "number"
                                                },
                                                "percentage": {
                                                    "type": "number"
                                                },
                                                "resetDate": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "subscriptionTier": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/models": {
            "get": {
                "tags": [
                    "Models"
                ],
                "summary": "List Models",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of models",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "image_url": {
                                                        "type": "string"
                                                    },
                                                    "is_public": {
                                                        "type": "boolean"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Models"
                ],
                "summary": "Create Model",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "name",
                                    "imageUrl"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "imageUrl": {
                                        "type": "string"
                                    },
                                    "subtitle": {
                                        "type": "string"
                                    },
                                    "projectId": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Model created"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Models"
                ],
                "summary": "Delete Model",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Model deleted"
                    }
                }
            }
        },
        "/locations": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "List Locations",
                "responses": {
                    "200": {
                        "description": "List of locations"
                    }
                }
            },
            "post": {
                "tags": [
                    "Locations"
                ],
                "summary": "Create Location",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "name",
                                    "imageUrl"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "imageUrl": {
                                        "type": "string"
                                    },
                                    "subtitle": {
                                        "type": "string"
                                    },
                                    "projectId": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Location created"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Locations"
                ],
                "summary": "Delete Location",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Location deleted"
                    }
                }
            }
        },
        "/camera-presets": {
            "get": {
                "tags": [
                    "Camera Settings"
                ],
                "summary": "List Camera Presets",
                "responses": {
                    "200": {
                        "description": "List of camera presets",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "presets": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "id": {
                                                                "type": "string"
                                                            },
                                                            "title": {
                                                                "type": "string"
                                                            },
                                                            "prompt": {
                                                                "type": "string"
                                                            },
                                                            "isBuiltIn": {
                                                                "type": "boolean"
                                                            }
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Camera Settings"
                ],
                "summary": "Create Camera Preset",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "id",
                                    "title",
                                    "prompt"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "prompt": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preset created"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Camera Settings"
                ],
                "summary": "Delete Camera Preset",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Preset deleted"
                    }
                }
            }
        },
        "/upload-image": {
            "post": {
                "tags": [
                    "Models"
                ],
                "summary": "Upload Image",
                "description": "Upload a file to R2. Use `skipDb=true` to get URL without DB insert. Shared utility for Models and Locations.",
                "parameters": [
                    {
                        "name": "skipDb",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "model",
                                            "location",
                                            "storyboard"
                                        ]
                                    },
                                    "projectId": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Upload successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "publicUrl": {
                                            "type": "string"
                                        },
                                        "key": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/generated-images": {
            "get": {
                "tags": [
                    "Gallery"
                ],
                "summary": "List Generated Images",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 24
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "favorites",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of generated images",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "image_url": {
                                                        "type": "string"
                                                    },
                                                    "prompt": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        },
                                        "hasMore": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}